feat(db): Add "naive" PostgreSQL implementation
Copy/paste of the SQLite implementation, and necessary changes
to SQL syntax and data types.

The rationale for choosing Postgres was the fact that `rusqlite`
crate was based on `postgres` crate, thus sharing similarities
in the API. They have, however, diverged significantly, so this
turned out to be not as much of advantage as was expected
initially.

Notable differences:
  - PostgreSQL does not have u64 type, and u32 is used for OIDs,
    so this implementation stores u64 as i64, and u32 as i32.
    On retrieval, these are converted back to uXX.
  - PostgreSQL does not have `INSERT OR REPLACE`, instead
    "upsert" syntax is used, which is `INSERT ON CONFLICT DO SET ...`.
  - PostgreSQL uses `$` to specify variables in the queries, instead
    of SQLite `:varname` syntax.
  - PostgreSQL types differ from SQLite (INT = i32, BIGINT = i64, ...).

Design decisions:
  - u64 and u32 types are converted to i64 and i32. It is assumed
    that no loss occurs by casting these types to and back.
  - Every test creates a new random database.

Known shortcomings:
  - At present, test databases are not deleted after tests have
    finished, and are removed manually via `psql` if required.